+++ src/main/java/nals/hrm/api_nals_hrm/service/LoginService.java
@@ -24,7 +24,7 @@ import java.util.List;
 
 @Service
 @Transactional
-public class EmployeeService {
+public class LoginService {
 
     @Autowired
     private EmployeeRepository employeeRepository;
