+++ src/main/resources/application.properties
@@ -2,7 +2,7 @@
 # ===============================
 # DATASOURCE
 # ===============================
-server.port=8081
+server.port=8082
 #Connection url for the database "nal_hrm"
 spring.datasource.url = jdbc:mysql://localhost:3306/nal_hrm?autoReconnect=true&useSSL=false&useUnicode=yes&characterEncoding=UTF-8
